requirements.txt
Django==5.1.6
django-cors-headers==4.7.0
djangorestframework==3.15.2
aiohttp==3.11.12
pydantic==2.10.6
huggingface-hub==0.28.1
groq==0.18.0
python-dotenv==1.0.1